Move commands to allow hot reloads
